telegram_bot_tamxu/requirements.txt
python-telegram-bot==22.3
gspread==6.0.2
oauth2client==4.1.3
web3==7.13.0
